refactor(llamalend): replace page overlay merge with canonical table data query
